Change internet_is_connected into an import
